pyproject.toml
[build-system]
requires = ["setuptools>=70", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "quizmaster"
version = "2.0.0"
description = "A modern question bank generator using qBank and BookWorm for intelligent content processing and adaptive learning"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Core frameworks
    "qbank @ git+https://github.com/haxx0rman/qBank.git",
    "bookworm",
    # LLM and AI
    "openai>=1.0.0",
    "anthropic>=0.34.0",
    "tiktoken>=0.5.0",
    # Core utilities
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "tqdm>=4.65.0",
    # Data handling
    "aiofiles>=23.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

[project.scripts]
quizmaster = "quizmaster.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["quizmaster*"]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.uv.sources]
bookworm = { git = "https://github.com/haxx0rman/BookWorm.git" }
